split loot dialog, added ui file
